exit if ctx is cancelled while waiting on read lock
git operations like `fetch` are slower and its called
under Mirror() which grabs write lock on repository.
Because of this functions like `Hash` can be waiting
for read lock for a long time and during that time
its given ctx can be cancelled.

If ctx is already cancelled then waiting on Rlock()
is waste of time since call to `runGitCommand(ctx, ...)` will fail anyway.